refactor : 정산 정보 저장 잡 리팩토링
