+++ DuDoong-Batch/src/main/java/band/gosrock/helper/excel/ExcelOrderDto.java
@@ -1,4 +1,4 @@
-package band.gosrock.excel;
+package band.gosrock.helper.excel;
 
 
 import band.gosrock.domain.common.vo.Money;
